refactor: Move shared enums to model folder
