+++ event-handlers/src/main/java/no/sikt/nva/nvi/events/evaluator/model/PointCalculation.java
@@ -3,8 +3,8 @@ package no.sikt.nva.nvi.events.evaluator.model;
 import java.math.BigDecimal;
 import java.net.URI;
 import java.util.List;
+import no.sikt.nva.nvi.common.model.InstanceType;
 import no.sikt.nva.nvi.common.model.ScientificValue;
-import no.sikt.nva.nvi.common.service.model.InstanceType;
 import no.sikt.nva.nvi.common.service.model.InstitutionPoints;
 
 public record PointCalculation(
