+++ event-handlers/src/test/java/no/sikt/nva/nvi/events/cristin/CristinMapperTest.java
@@ -29,8 +29,8 @@ import java.net.URI;
 import java.nio.file.Path;
 import java.util.List;
 import no.sikt.nva.nvi.common.db.CandidateDao.DbLevel;
-import no.sikt.nva.nvi.common.db.model.ChannelType;
 import no.sikt.nva.nvi.common.dto.PublicationDateDto;
+import no.sikt.nva.nvi.common.model.ChannelType;
 import no.sikt.nva.nvi.events.cristin.CristinNviReport.Builder;
 import no.unit.nva.commons.json.JsonUtils;
 import nva.commons.core.ioutils.IoUtils;
